Intermediate Commit to Update TestRunner.
P.S.: I already got another version prepared locally that should be able to do way more in a much more flexible way and be EVEN EASIER to use.

but it will still take me a while to finish that.

Until the new changes will be completely in here the TestRunner3 will do this now:

   if (attribute.GetType() != typeof(TestAttribute))
                            {
                                var handler = AttributeLogicMapper.GetHandlerFor(attribute);
                                handler?.ProcessAttribute(attribute, method, classObject);
                                if (handler != null)
                                {
                                    testResult.State = ExecutedState.Passed;
                                }
                            }
                            else if (definition.Attributes.Count() == 1 && attribute is TestAttribute)
                            {
                                method.Invoke(classObject, null);
                                testResult.State = ExecutedState.Passed;
                            }

To check for execution and then finally execute, if marked correctly.

The AattributeLogicMapper got a new method to get the handlers (method signature below):

public IAttributeLogicHandler GetHandlersFor((Type type, MethodInfo methodInfo, IEnumerable<Attribute> Attributes) definition, object classObject)

This should make it even much easier in the future and more sophisticated to get the correct handler. And can also be used with the combinator later on.

That's it for now.