First check in to the new Development branch.
NextUnit.AssembyReader
---------------------------
We got the new AssemblyReader which contains now a PDBReader. What is this used for?
Well, since I found out that it may not be THAT straightforward to show the user WHERE your code failed we have to determine the source and the lines of code, so... You know? ;-)

As well as a ReflectionExtensions.GetSources has been added to get all source files of the assembly. Ideally we have then only to get ALL assemblies, get the source files and then parse.

NextUnit.Fuzzer
-----------------
The fuzzer will help contain the logic for the FuzzAttribute.
Possibly this may become a fuzzing library itself. BUT for sure not an advanced fuzzing framework. More leveraging what exists and provide some basics.

FileSourcesParser
-------------------
Belongs to the same function category as the PDBReader. This will be used to read files to show where something crashed/failed, etc.

NextUnit.InstantTests
----------------------
This will become a mechanism to instantly recompile/run tests as specified/configured.
For now the following is thinkable:

A console runner will be started OR a visual studio Test Explorer Run will be started.

1. And every test will be run.
2. Only the affected tests will be run.

NextUnit.VisualStudioExtension
----------------------------------
Should become a VisualStudio Extension to run tests as easy a using resharper for instance.

NextUnit.VisualStudioIntegration
-----------------------------------
Will be used by VisualStudioExtension or possibly InstantTests to integrate this to visual studio.

PDBReader got some code to be finally able to detect code files.

NextUnit.Core
---------------
The FuzzingAttributeLogicHandler has been rewritten.

NextUnit.TestRunner
----------------------
- Combinators can now be added at runtime (checks for the allowed value combinations and generates TestRuns out of them)
  Example: The DefaultCombinator will now resemble the same behavior as in previous versions: Just run as before.
- A new combinator will be responsible to run everything "correctly", which means pay attention to combinations and create TestRuns based on them.
- InstanceCreationBehaviors can now also be set even during runtime. If IInstanceCreationBehavior is RecreateObjectInstanceForEachTest then it will work as in XUnit.
  This means if you would write something like:
   public class TestClass
   {
       Guid guid = new Guid();
   }
   you will get a new Guid for each ran TestMethod.

  If we set it to CacheInstanceObjects the objects will be taken from a dictionary and not be recreated. Thus, Guid in the above example will always be the same
- The TestDiscoverer uses now a better way to discover TestMethods. Way easier and way more reliable. This is working very well.
- The TestResult now also has a CreationTime in there. This will be good to track to see when the TestResult has been created and when the test was run or not. Very helpful for debugging purposes.
- The TestRunnerAssemblyLoadContext can now be called from outside.
- The NextUnitBaseExecutor uses now the ITestRunner3 interface instead of the TestRunner3 object instance.
- The TestDiscoverer now shows the traits correctly and the fullyQualifiedName.

NextUnit.Console.TestRunner
- The TestRunner has been exchanged and uses now TestRunner4.