Some unneccessary example files have been removed.
Examples\Example:

- A test project has been added to show how to use the test runner.

NextUnit.AssemblyReader

- ReferenceReader has been added.

NextUnit.Autofixture.AutoMoq.Core

- Every file in this folder or subfolders has been added to be able to use the new invoker logic.

NextUnit.Core

- dito

NextUnit.Core.InvokingMechanism

- different mechanisms for invoking have been added

KnownFolders has been extended to support more folders.
And SysInfo has therefore been adapted as well.

In TestAdapter some more settings that also could be set when watching Microsoft's projects have been added.
See DiscoveryContext, TestRunSettings, FrameworkHandle, etc. and/or adapted.

Tools
==========

Test Explorer
---------------
There's a first one theoretically fully usable TestExplorerUI, but it needs TestAdapter files (either of NextUnit or something else) in a special subfolder called "TestAdapter". Then it should be working.

The ConsoleExtensions of the TestRunner went to the core. Because it might be useful to use them in other places. When separating more they should go into another assembly.

ConsoleTools could be used to generate test stubs by selecting an assembly.

CreateTestSubsFromImplementedClassProofOfConcept

This can be used to - when you got a source class in your clipboard - to just run this program
then all the test stubs should be generated correctly and copied back to the clipboard.

For this do the following:

1. As said before: Copy your class into the clipboard.
2. Start the program.
3. Paste the test stubs in a file in Visual Studio.

That's how it works for now. The program has to be restarted for each generation.
But I will remove this issue for sure in a further version. Because it's quite annoying.
Maybe it will fuse with the Test Completeness UI which seems to make a lot of sense so far.

GetVersionsFromCSProjsProofOfConcept

Has the capability to show which assembly versions (DLL references, NuGet package references, etc.) are being used.
Thus, it would be way easier to figure out mismatches in your project.

TestCompletenessUI

Is a tool to show you when selecting a source file (*.cs c# file) and test file (*.cs c# file)
to see how many tests have already been written and how much is still missing.

This still is very error prone but a very good start. Because it doesn't just check by test names like one would assume, no.
It checks for calls to the function, i.e. if a the source class contains a method called "GetFiles" then it won't just look for
"GetFilesTest". No, it detects if a "GetFiles" call happens from one of the tests. Which is much more reliable.

Tests
------
Some tests had to be adapted for sure as well.
Test completeness is not 100% guaranteed in this version. But it has improved. Will be tested in a further version to see if it really always hits all the tests.